requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.2
boto3==1.34.69 